requirements.txt
pandas==1.5.3
numpy==1.24.2
PyYAML==6.0
torch==1.12.0
torchtext==0.14.1
scikit-learn==1.2.1
plotly==5.13.0
